MIM-1996: Refaktorering, flyttet kode til lib for gjenbruk
